restructure: layouts-pages 폴더 정리
